Revert SampleImageAnalysis id attribute to type Integer

+++ uk.ac.diamond.ispyb.jdbc/src/uk/ac/diamond/ispyb/dao/IspybPlateDAO.java
@@ -88,8 +88,8 @@ public class IspybPlateDAO implements IspybPlateApi{
 	}
 
 	@Override
-	public Long upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis){
-		return beanTemplateWrapper.callIspyb("upsert_sample_image_analysis", Long.class, sampleImageAnalysis).get();
+	public Integer upsertSampleImageAnalysis(SampleImageAnalysis sampleImageAnalysis){
+		return beanTemplateWrapper.callIspyb("upsert_sample_image_analysis", Integer.class, sampleImageAnalysis).get();
 	}
 	
 	@Override
